+++ app/Http/Controllers/Web/TypeWebController.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\Controller;
 use App\Http\Requests\TypeRequest;
 use App\Models\Type;
 use App\Repositories\TypeRepositoryInterface;
+use App\Support\ReferenceSlugs;
 
 class TypeWebController extends Controller
 {
@@ -20,27 +21,33 @@ class TypeWebController extends Controller
 
     public function create()
     {
-        return view('settings.types.create');
+        abort(404);
     }
 
     public function store(TypeRequest $request)
     {
-        $data = $request->validated();
-
-        $this->types->create($data);
-
-        return redirect()
-            ->route('types.index')
-            ->with('success', 'Tipo criado com sucesso!');
+        abort(404);
     }
 
     public function edit(Type $type)
     {
+        if (ReferenceSlugs::isSystemTypeSlug($type->slug)) {
+            return redirect()
+                ->route('types.index')
+                ->withErrors(['type' => 'Este tipo é usado pelo sistema e não pode ser alterado.']);
+        }
+
         return view('settings.types.edit', compact('type'));
     }
 
     public function update(TypeRequest $request, Type $type)
     {
+        if (ReferenceSlugs::isSystemTypeSlug($type->slug)) {
+            return redirect()
+                ->route('types.index')
+                ->withErrors(['type' => 'Este tipo é usado pelo sistema e não pode ser alterado.']);
+        }
+
         $data = $request->validated();
 
         $this->types->update($type->id, $data);
@@ -52,11 +59,16 @@ class TypeWebController extends Controller
 
     public function destroy(Type $type)
     {
+        if (ReferenceSlugs::isSystemTypeSlug($type->slug)) {
+            return redirect()
+                ->route('types.index')
+                ->withErrors(['type' => 'Este tipo é usado pelo sistema e não pode ser excluído.']);
+        }
+
         $this->types->delete($type->id);
 
         return redirect()
             ->route('types.index')
             ->with('success', 'Tipo removido com sucesso!');
     }
-
 }
